extract diStorage in module
