+++ app/build.gradle
@@ -37,6 +37,8 @@ android {
 }
 
 dependencies {
+    implementation project(path: ':dilib')
+
     implementation 'androidx.room:room-runtime:2.4.2'
     kapt 'androidx.room:room-compiler:2.4.2'
     implementation 'androidx.room:room-rxjava3:2.4.2'
